\name{IPMMc}
\alias{IPMMc}
\docType{data}

\title{Inventory of Personality and Mood Manifestation} 

\description{
The \code{IPMMc} data frame has 6 rows and 4 columns. See Davison & Davenport (2002) for more information.  
}

\format{
  This data frame contains the following columns:
  \describe{
    \item{A}{Anxiety}
    \item{H}{Hypochondriasis}
    \item{S}{Schizophrenia}
    \item{B}{Bipolar Disorder}
   }
 }

\source{
Davison, M., & Davenport, E. (2002). Identifying criterion-related patterns of predictor scores using multiple regression. Psychological Methods, 7(4), 468-484.

}

\references{
Davison, M., & Davenport, E. (2002). Identifying criterion-related patterns of predictor scores using multiple regression. Psychological Methods, 7(4), 468-484.
}

\usage{
IPMMc
}

\keyword{datasets}
