\name{NP}
\alias{NP}
\docType{data}

\title{The Neurotic versus Psychotic Criterion Variable}

\description{
The \code{NP} data frame has 6 rows and 1 column. See Davison & Davenport (2002) for more information.  
}

\format{
  This data frame contains the following columns:
  \describe{
    \item{NP}{Neurotic (= 1) or Psychotic (= 0)}
   }
 }

\source{
Davison, M., & Davenport, E. (2002). Identifying criterion-related patterns of predictor scores using multiple regression. Psychological Methods, 7(4), 468-484.

}

\references{
Davison, M., & Davenport, E. (2002). Identifying criterion-related patterns of predictor scores using multiple regression. Psychological Methods, 7(4), 468-484.
}

\usage{
NP
}

\keyword{datasets}
