\name{profanal-package}
\alias{profanal-package}
\alias{profanal}
\docType{package}
\title{
Profile Analysis
}
\description{
Implements the profile analysis and cross-validation techniques described in Davison & Davenport (2002).
}
\details{
\tabular{ll}{
Package: \tab profanal\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-02-25\cr
License: \tab GPL-3\cr
}

Index:
\preformatted{
IPMMc                   Inventory of Personality and Mood Manifestations
NP                   	Neurotic or Psychotic Diagnosis
criterion.pattern       Performs the Criterion-Related Profile Analysis
profilecv              Performs the Cross-Validation Technique
}

Further information is available in the following vignette:
\tabular{ll}{
\code{profanalnotes} \tab profanalnotes vignette (source, pdf)\cr
}
The package vignette provides an introduction to the criterion-related profile analysis described in Davison & Davenport. The help page for the main function of the package, \code{\link{profanal}}, contains additional examples.
}

\author{
Christopher David Desjardins

Maintainer: Christopher David Desjardins <cddesjardins@gmail.com>
}
\references{
Davison, M., & Davenport, E. (2002). Identifying criterion-related patterns of predictor scores using multiple regression. Psychological Methods, 7(4), 468-484.
}
\keyword{package}